pyproject.toml
[project]
name = "mumbaihacks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepgram-sdk==3.4.0",
    "fastapi>=0.122.0",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=3.2.0",
    "langchain-groq>=1.1.0",
    "langchain-openai>=1.1.0",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "supabase>=2.24.0",
    "tiktoken>=0.12.0",
    "uvicorn>=0.38.0",
    "websockets>=14.0",
]
